Extract channel cell to separate component
